Merge and refactor changes made by Filipe regarding stopwords processing. Instead of using a hardcoded list of stopwords in the Java class, we read them all from files. That will make the update process easier (and also free us from Java's 64k limit for compiled class size).
Several changes were also made regarding parsing and writing files.
There are two methods to that: one to read data and other to write
date. The major change, actually, is regarding the correctness of the
implementation: we are now fully compatible with DTM format. The old
approach was broken: documents were not ordered by date and terms for
each document were being incorrectly accounted. Terms are now also
identified by numbers from 1 to n (thus compatible with the assumption
that the line number is the term identifier).

Furthermore, the program now generates two extra files: one with terms
and another with documents. Hopefully, we will be able to use them
to fully understand the data DTM provides us.